requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile
#
certifi==2021.10.8
    # via requests
charset-normalizer==2.0.12
    # via requests
idna==3.3
    # via requests
numpy==1.22.2
    # via -r requirements.in
parse==1.19.0
    # via -r requirements.in
pyyaml==6.0
    # via -r requirements.in
requests==2.27.1
    # via -r requirements.in
urllib3==1.26.8
    # via requests
